Incorporate main-only changes to resolve dev-to-main conflicts
Bring forward changes that landed on main but not dev:
- CHANGELOG version renumbering (0.77-0.91 bumped to 0.78-0.92) and
  new 0.77.0 entry for frontend component extraction
- Frontend component extraction: split graph, requirement-detail, and
  requirements pages into focused modules with route-level code splitting
- Baseline cross-project comparison fix (DomainValidationException)
  with controller and service tests
