+++ frontend/src/components/requirements/filter-select.tsx
@@ -1,0 +1,26 @@
+export function FilterSelect({
+  value,
+  onChange,
+  placeholder,
+  options,
+}: {
+  value: string;
+  onChange: (v: string) => void;
+  placeholder: string;
+  options: string[];
+}) {
+  return (
+    <select
+      className="rounded-md border border-input bg-background px-2 py-1.5 text-sm text-foreground"
+      value={value}
+      onChange={(e) => onChange(e.target.value)}
+    >
+      <option value="">{placeholder}</option>
+      {options.map((o) => (
+        <option key={o} value={o}>
+          {o.replace(/_/g, " ")}
+        </option>
+      ))}
+    </select>
+  );
+}
