+++ frontend/src/routes.tsx
@@ -1,14 +1,40 @@
 import { AppLayout } from "@/components/layout/app-layout";
 import { ProjectProvider } from "@/contexts/project-context";
 import { useProjects } from "@/hooks/use-projects";
-import { Admin } from "@/pages/admin";
-import { Analysis } from "@/pages/analysis";
-import { Dashboard } from "@/pages/dashboard";
-import { Graph } from "@/pages/graph";
-import { Projects } from "@/pages/projects";
-import { RequirementDetail } from "@/pages/requirement-detail";
-import { Requirements } from "@/pages/requirements";
+import { Suspense, lazy } from "react";
 import { Link, Navigate, Route, Routes } from "react-router-dom";
+
+const Admin = lazy(() =>
+  import("@/pages/admin").then((m) => ({ default: m.Admin })),
+);
+const Analysis = lazy(() =>
+  import("@/pages/analysis").then((m) => ({ default: m.Analysis })),
+);
+const Dashboard = lazy(() =>
+  import("@/pages/dashboard").then((m) => ({ default: m.Dashboard })),
+);
+const Graph = lazy(() =>
+  import("@/pages/graph").then((m) => ({ default: m.Graph })),
+);
+const Projects = lazy(() =>
+  import("@/pages/projects").then((m) => ({ default: m.Projects })),
+);
+const RequirementDetail = lazy(() =>
+  import("@/pages/requirement-detail").then((m) => ({
+    default: m.RequirementDetail,
+  })),
+);
+const Requirements = lazy(() =>
+  import("@/pages/requirements").then((m) => ({ default: m.Requirements })),
+);
+
+function PageSkeleton() {
+  return (
+    <div className="flex min-h-[60vh] items-center justify-center">
+      <div className="h-8 w-8 animate-spin rounded-full border-4 border-muted border-t-primary" />
+    </div>
+  );
+}
 
 function NotFound() {
   return (
@@ -42,30 +68,32 @@ function RootRedirect() {
 
 export function AppRoutes() {
   return (
-    <Routes>
-      <Route path="/" element={<RootRedirect />} />
-      <Route element={<AppLayout />}>
-        <Route path="projects" element={<Projects />} />
-      </Route>
-      <Route
-        path="p/:projectId"
-        element={
-          <ProjectProvider>
-            <AppLayout />
-          </ProjectProvider>
-        }
-      >
-        <Route index element={<Dashboard />} />
-        <Route path="requirements" element={<Requirements />} />
-        <Route path="requirements/:id" element={<RequirementDetail />} />
-        <Route path="graph" element={<Graph />} />
-        <Route path="analysis" element={<Analysis />} />
-        <Route path="admin" element={<Admin />} />
-        <Route path="*" element={<NotFound />} />
-      </Route>
-      <Route element={<AppLayout />}>
-        <Route path="*" element={<NotFound />} />
-      </Route>
-    </Routes>
+    <Suspense fallback={<PageSkeleton />}>
+      <Routes>
+        <Route path="/" element={<RootRedirect />} />
+        <Route element={<AppLayout />}>
+          <Route path="projects" element={<Projects />} />
+        </Route>
+        <Route
+          path="p/:projectId"
+          element={
+            <ProjectProvider>
+              <AppLayout />
+            </ProjectProvider>
+          }
+        >
+          <Route index element={<Dashboard />} />
+          <Route path="requirements" element={<Requirements />} />
+          <Route path="requirements/:id" element={<RequirementDetail />} />
+          <Route path="graph" element={<Graph />} />
+          <Route path="analysis" element={<Analysis />} />
+          <Route path="admin" element={<Admin />} />
+          <Route path="*" element={<NotFound />} />
+        </Route>
+        <Route element={<AppLayout />}>
+          <Route path="*" element={<NotFound />} />
+        </Route>
+      </Routes>
+    </Suspense>
   );
 }
